master merge fixes
