+++ packages/desktop-client/src/components/reports/spreadsheets/grouped-spreadsheet.ts
@@ -11,18 +11,22 @@ import makeQuery from './makeQuery';
 import recalculate from './recalculate';
 
 function createGroupedSpreadsheet({
-  start,
-  end,
+  startDate,
+  endDate,
   categories,
   selectedCategories,
   conditions = [],
   conditionsOp,
-  hidden,
-  uncat,
+  showOffBudgetHidden,
+  showUncategorized,
 }: createSpreadsheetProps) {
-  const [catList, catGroup] = categoryLists(hidden, uncat, categories);
+  const [categoryList, categoryGroup] = categoryLists(
+    showOffBudgetHidden,
+    showUncategorized,
+    categories,
+  );
 
-  const categoryFilter = (catList || []).filter(
+  const categoryFilter = (categoryList || []).filter(
     category =>
       !category.hidden &&
       selectedCategories &&
@@ -32,7 +36,7 @@ function createGroupedSpreadsheet({
   );
 
   return async (spreadsheet, setData) => {
-    if (catList.length === 0) {
+    if (categoryList.length === 0) {
       return null;
     }
 
@@ -45,9 +49,9 @@ function createGroupedSpreadsheet({
       runQuery(
         makeQuery(
           'assets',
-          start,
-          end,
-          hidden,
+          startDate,
+          endDate,
+          showOffBudgetHidden,
           selectedCategories,
           categoryFilter,
           conditionsOpKey,
@@ -57,9 +61,9 @@ function createGroupedSpreadsheet({
       runQuery(
         makeQuery(
           'debts',
-          start,
-          end,
-          hidden,
+          startDate,
+          endDate,
+          showOffBudgetHidden,
           selectedCategories,
           categoryFilter,
           conditionsOpKey,
@@ -68,71 +72,73 @@ function createGroupedSpreadsheet({
       ).then(({ data }) => data),
     ]);
 
-    const months = monthUtils.rangeInclusive(start, end);
+    const months = monthUtils.rangeInclusive(startDate, endDate);
 
-    const groupedData = catGroup.map(
-      group => {
-        let totalAssets = 0;
-        let totalDebts = 0;
+    const groupedData = categoryGroup
+      .filter(f => (showOffBudgetHidden || f.hidden === false) && f)
+      .map(
+        group => {
+          let totalAssets = 0;
+          let totalDebts = 0;
 
-        const monthData = months.reduce((arr, month) => {
-          let groupedAssets = 0;
-          let groupedDebts = 0;
+          const monthData = months.reduce((arr, month) => {
+            let groupedAssets = 0;
+            let groupedDebts = 0;
 
-          group.categories.map(item => {
-            const monthAssets = filterHiddenItems(item, assets)
-              .filter(
-                asset => asset.date === month && asset.category === item.id,
-              )
-              .reduce((a, v) => (a = a + v.amount), 0);
-            groupedAssets += monthAssets;
+            group.categories.map(item => {
+              const monthAssets = filterHiddenItems(item, assets)
+                .filter(
+                  asset => asset.date === month && asset.category === item.id,
+                )
+                .reduce((a, v) => (a = a + v.amount), 0);
+              groupedAssets += monthAssets;
 
-            const monthDebts = filterHiddenItems(item, debts)
-              .filter(
-                debts => debts.date === month && debts.category === item.id,
-              )
-              .reduce((a, v) => (a = a + v.amount), 0);
-            groupedDebts += monthDebts;
+              const monthDebts = filterHiddenItems(item, debts)
+                .filter(
+                  debts => debts.date === month && debts.category === item.id,
+                )
+                .reduce((a, v) => (a = a + v.amount), 0);
+              groupedDebts += monthDebts;
 
-            return null;
+              return null;
+            });
+
+            totalAssets += groupedAssets;
+            totalDebts += groupedDebts;
+
+            arr.push({
+              date: month,
+              totalAssets: integerToAmount(groupedAssets),
+              totalDebts: integerToAmount(groupedDebts),
+              totalTotals: integerToAmount(groupedDebts + groupedAssets),
+            });
+
+            return arr;
+          }, []);
+
+          const stackedCategories = group.categories.map(item => {
+            const calc = recalculate({
+              item,
+              months,
+              assets,
+              debts,
+              groupByLabel: 'category',
+            });
+            return { ...calc };
           });
 
-          totalAssets += groupedAssets;
-          totalDebts += groupedDebts;
-
-          arr.push({
-            date: month,
-            totalAssets: integerToAmount(groupedAssets),
-            totalDebts: integerToAmount(groupedDebts),
-            totalTotals: integerToAmount(groupedDebts + groupedAssets),
-          });
-
-          return arr;
-        }, []);
-
-        const stackedCategories = group.categories.map(item => {
-          const calc = recalculate({
-            item,
-            months,
-            assets,
-            debts,
-            groupByLabel: 'category',
-          });
-          return { ...calc };
-        });
-
-        return {
-          id: group.id,
-          name: group.name,
-          totalAssets: integerToAmount(totalAssets),
-          totalDebts: integerToAmount(totalDebts),
-          totalTotals: integerToAmount(totalAssets + totalDebts),
-          monthData,
-          categories: stackedCategories,
-        };
-      },
-      [start, end],
-    );
+          return {
+            id: group.id,
+            name: group.name,
+            totalAssets: integerToAmount(totalAssets),
+            totalDebts: integerToAmount(totalDebts),
+            totalTotals: integerToAmount(totalAssets + totalDebts),
+            monthData,
+            categories: stackedCategories,
+          };
+        },
+        [startDate, endDate],
+      );
 
     setData(groupedData);
   };
